v240

+++ src/eu/engys/core/executor/CommandExecutor.java
@@ -1,33 +1,33 @@
-/*--------------------------------*- Java -*---------------------------------*\
- |		 o                                                                   |                                                                                     
- |    o     o       | HelyxOS: The Open Source GUI for OpenFOAM              |
- |   o   O   o      | Copyright (C) 2012-2016 ENGYS                          |
- |    o     o       | http://www.engys.com                                   |
- |       o          |                                                        |
- |---------------------------------------------------------------------------|
- |	 License                                                                 |
- |   This file is part of HelyxOS.                                           |
- |                                                                           |
- |   HelyxOS is free software; you can redistribute it and/or modify it      |
- |   under the terms of the GNU General Public License as published by the   |
- |   Free Software Foundation; either version 2 of the License, or (at your  |
- |   option) any later version.                                              |
- |                                                                           |
- |   HelyxOS is distributed in the hope that it will be useful, but WITHOUT  |
- |   ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or   |
- |   FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License   |
- |   for more details.                                                       |
- |                                                                           |
- |   You should have received a copy of the GNU General Public License       |
- |   along with HelyxOS; if not, write to the Free Software Foundation,      |
- |   Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA            |
-\*---------------------------------------------------------------------------*/
-
+/*******************************************************************************
+ *  |       o                                                                   |
+ *  |    o     o       | HELYX-OS: The Open Source GUI for OpenFOAM             |
+ *  |   o   O   o      | Copyright (C) 2012-2016 ENGYS                          |
+ *  |    o     o       | http://www.engys.com                                   |
+ *  |       o          |                                                        |
+ *  |---------------------------------------------------------------------------|
+ *  |   License                                                                 |
+ *  |   This file is part of HELYX-OS.                                          |
+ *  |                                                                           |
+ *  |   HELYX-OS is free software; you can redistribute it and/or modify it     |
+ *  |   under the terms of the GNU General Public License as published by the   |
+ *  |   Free Software Foundation; either version 2 of the License, or (at your  |
+ *  |   option) any later version.                                              |
+ *  |                                                                           |
+ *  |   HELYX-OS is distributed in the hope that it will be useful, but WITHOUT |
+ *  |   ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or   |
+ *  |   FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License   |
+ *  |   for more details.                                                       |
+ *  |                                                                           |
+ *  |   You should have received a copy of the GNU General Public License       |
+ *  |   along with HELYX-OS; if not, write to the Free Software Foundation,     |
+ *  |   Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA            |
+ *******************************************************************************/
 package eu.engys.core.executor;
 
 import static eu.engys.core.OpenFOAMEnvironment.cleanEnvironment;
 import static eu.engys.core.OpenFOAMEnvironment.loadEnvironment;
 import static eu.engys.core.OpenFOAMEnvironment.printHeader;
+import static eu.engys.core.OpenFOAMEnvironment.printVariables;
 
 import java.io.File;
 import java.util.List;
@@ -58,7 +58,11 @@ public class CommandExecutor extends AbstractScriptExecutor {
 
     @Override
     protected void internalDeleteOnEnd() {
-        FileUtils.deleteQuietly(supportFile);
+        if(keepFileOnEnd){
+            //do nothing
+        } else {
+            FileUtils.deleteQuietly(supportFile);
+        }
     }
 
     private void writeCommandInSupportFile(File supportFile) {
@@ -69,6 +73,7 @@ public class CommandExecutor extends AbstractScriptExecutor {
     private List<String> getCommand() {
         ScriptBuilder sb = new ScriptBuilder();
         printHeader(sb, description);
+        printVariables(sb);
         if (Util.isWindows()) {
             if (loadOpenFoamEnv) {
                 loadEnvironment(sb);
